requirements/requirements.txt
# Основные зависимости
numpy>=1.21.0
pandas>=1.3.0
networkx>=2.6.0
tqdm>=4.62.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Базы данных
psycopg2-binary>=2.9.0
neo4j>=5.0.0
python-dateutil>=2.8.0

# Утилиты
pyyaml>=6.0
jupyter>=1.0.0
ipython>=8.0.0

# Для статистического анализа
scipy>=1.7.0
scikit-learn>=1.0.0
